separate modules from models + BrainLargeDataset __init__
