+++ scvi/models/__init__.py
@@ -1,0 +1,5 @@
+from .vae import VAE
+from .vaec import VAEC
+
+__all__ = ['VAE',
+           'VAEC']
